@@ -2,15 +2,16 @@
   "name": "bulletinboard-ads",
   "version": "1.0.0",
   "description": "",
-  "main": "main.js",
+  "main": "js/index.js",
+  "type": "module",
   "scripts": {
-    "start": "node js/main.js",
-    "start:debug": "node --inspect-brk js/main.js",
+    "start": "node js/index.js",
+    "start:debug": "node --inspect-brk js/index.js",
     "test": "node node_modules/mocha/bin/mocha",
     "test:debug": "node node_modules/mocha/bin/mocha --inspect-brk"
   },
   "engines": {
-    "node": "12.16.x"
+    "node": "16.12.x"
   },
   "author": "",
   "license": "ISC",
@@ -21,7 +22,7 @@
     "winston": "^3.3.3"
   },
   "devDependencies": {
-    "mocha": "^9.1.2",
+    "mocha": "^9.1.3",
     "sinon": "^11.1.2",
     "supertest": "^6.1.6"
   }
