@@ -1,9 +1,8 @@
-'use strict'
-const winston = require('winston')
+import winston  from 'winston'
 
 const create = () => {
     const logFormat = winston.format.combine(winston.format.splat(), winston.format.simple())
     return winston.createLogger({ transports: [new winston.transports.Console()], format: logFormat })
 }
 
-module.exports = { create }
+export default { create }
